requirements.txt
Pillow
opencv-python
numpy
scipy
matplotlib
docopt